# contracts/openapi/apgic-v1.yaml
openapi: 3.1.0
info:
  title: APGIC Public API
  version: 0.10.0-r0-remote-config
servers: []
paths:
...
  /v1/mobile/checkout-instructions:
    post:
      operationId: createMobileCheckoutInstruction
      description: Executes the canonical checkout command through the same demand/domain truth as web while binding mobile timeout/retry to one server-side idempotency record.
      security:
        - ClientSession: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - in: header
          name: X-Correlation-Id
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileCheckoutMutationRequest'
      responses:
        '201':
          description: Canonical checkout side effect applied and acknowledged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileCheckoutMutationResponse'
        '200':
          description: Idempotent retry resolved to the already applied canonical checkout side effect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileCheckoutMutationResponse'
        '400':
          description: Invalid mutation envelope or checkout input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Trusted client session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Idempotency conflict, terminal prior failure or canonical checkout conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Mutation persistence or canonical checkout runtime is temporarily unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    ClientSession:
      type: apiKey
      in: cookie
      name: __Host-apgic_session
      description: Server-signed trusted client session. Caller-supplied identity fields are never an authorization source.
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [code, message_safe, correlation_id, retryable]
      properties:
        code: { type: string }
        message_safe: { type: string }
        correlation_id: { type: string }
        retryable: { type: boolean }
        field_errors:
          type: array
          items: { type: object }
        policy_reason_codes:
          type: array
          items: { type: string }
        documentation_ref: { type: string }
...
    MobileCheckoutMutationRequest:
      type: object
      additionalProperties: false
      required: [hold_id, method_code]
      properties:
        hold_id: { type: string, minLength: 1 }
        method_code: { type: string, minLength: 1 }
    MobileCheckoutMutationResponse:
      type: object
      additionalProperties: false
      required: [contract_version, outcome, reason_code, side_effect_ref, checkout]
      properties:
        contract_version:
          type: string
          enum: [offline-checkout-mutation-v1]
        outcome:
          type: string
          enum: [APPLIED, DUPLICATE_APPLIED]
        reason_code: { type: string, minLength: 1 }
        side_effect_ref: { type: string, minLength: 1 }
        checkout:
          $ref: '#/components/schemas/CheckoutInstruction'
...
    CheckoutInstruction:
      type: object
      additionalProperties: false
      required: [id, hold_id, booking_id, booking_state, order_id, provider_id, method_code, rail_code, amount_minor, currency, execution_owner, payment_recipient_id, platform_role, apgic_accepts_funds, notice, reason_code]
      properties:
        id: { type: string }
        hold_id: { type: string }
        booking_id: { type: string }
        booking_state: { type: string }
        order_id: { type: string }
        provider_id: { type: string }
        method_code: { type: string }
        rail_code: { type: string }
        amount_minor: { type: integer }
        currency: { type: string }
        execution_owner: { type: string }
        payment_recipient_id: { type: string }
        platform_role: { type: string }
        apgic_accepts_funds: { type: boolean }
        notice: { type: string }
        reason_code: { type: string }
...